Move config-changed to backup manager
